cdk-tgw-outbound/master/requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-ram
aws-cdk.aws-ssm